Dynamic Comments, OKS_Mechanized Hotfix?, Paradrop Reinsert JIP fix?
Dynamic comments are now available in Eden, shows what the different locations do for the dynamic scripts.

OKS_Mechanized, added a sleep to try and stop the FW from deleting the equipment added by the code. Hopefully should help with the Ammo issue previously had in Bulldog Missions.

Paradrop Reinsert - Is now disabled in missionSetting.sqf and by deleting the trigger for it. Hopefully this should work for JIP now.