Moderate Change: Gear Fix for AI not receiving correct kits.
Gear fix for AI not receiving correct kits. Our local handler is now disabled as it wasn't working properly and we now rely on blacklisting the units from transfer to HC until we have given them the correct gear. GearApplied variable is now globally set to make sure the HC also knows when the gear is applied.